% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionCal.R
\name{conservityTable}
\alias{conservityTable}
\title{This function calculates the conservity of all positions for input
sequences and generate a vector contains each position's conservity}
\usage{
conservityTable(seqSet)
}
\arguments{
\item{seqSet}{seqSet A set of aligned sequence read from fasta file}
}
\value{
a vector contains each position's
}
\description{
This function calculates the conservity of all positions for input
sequences and generate a vector contains each position's conservity
}
