% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionCal.R
\name{createPosVec}
\alias{createPosVec}
\title{Find most conserved nucleotide in each position and its frequency}
\usage{
createPosVec(seqSet, index)
}
\arguments{
\item{seqSet}{A set of aligned sequence read from fasta file}

\item{index}{A numeric index for generating Position String at that position.}
}
\value{
A String containing nucleotides at index position for sequences.
}
\description{
This function grab the nucleotide for all of sequences at index position
to create a XString sequence for a collection of this nucleotide. Then
the function reformats the Xstring into a string clean it up.
}
\examples{
# Example 1: Use a sequence set with short sequence from data,
# create a string containing nucleotides at index 1 for this sequence set.

data(testSeqShort)
testSeq <- testSeqShort
position1Vec <- createPosVec(testSeq, 1)
position1Vec # "ATT"

# Example 2: Use a raw data sequence set, create a string containing
# nucleotides at index 10 for these raw data sequences.

rawSeq <- Biostrings::readBStringSet("./inst/extdata/sampleSeq3.txt","fasta")
position10Vec <- createPosVec(rawSeq, 10)
position10Vec <- "TTTCACTACCAA"
}
