% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionCal.R
\name{findConservedNul}
\alias{findConservedNul}
\title{This function finds the nucleotide that has highest frequency of
in input string.}
\usage{
findConservedNul(str)
}
\arguments{
\item{str}{A string containing character nucleotides}
}
\value{
Return a vector that has one or more nucleotides
         with highest frequency
}
\description{
This function finds the nucleotide that has highest frequency of
in input string.
}
