% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionCal.R
\name{findConservityFromS}
\alias{findConservityFromS}
\title{This function finds the highest frequency of nucleotide in input string.}
\usage{
findConservityFromS(str)
}
\arguments{
\item{str}{A string containing character nucleotides}
}
\value{
Return a numeric value for highest frequency
}
\description{
This function finds the highest frequency of nucleotide in input string.
}
\examples{
# Example 1: find the highest letter frequency in a string

str <- "AADDDCCEEEE"
freq <- findConservityFromS(str)
freq # 0.363636...

# Example 2: find the highest frequency of nucleotide from a string of
# nucleotide generated by createPosVec()
data(testSeqShort)
position1Vec <- createPosVec(testSeqShort, 1)
freq <- findConservityFromS(position1Vec)
freq # 0.66666...
}
