% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionCal.R
\name{findMaxLen}
\alias{findMaxLen}
\title{This function finds the max length of sequence for input sequence set}
\usage{
findMaxLen(seqSet)
}
\arguments{
\item{seqSet}{seqSet A set of aligned sequence read from fasta file}
}
\value{
A numeric value of the max length of sequence for seqSet
}
\description{
This function finds the max length of sequence for input sequence set
}
