% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConser.R
\name{plotOverall}
\alias{plotOverall}
\title{Plot the conservity of the positions}
\usage{
plotOverall(conserV)
}
\arguments{
\item{conserV}{a vector with serveral number of conservity}
}
\value{
a plot to view the overall alignment's conservity
}
\description{
This function plots the overall alignment's conservity
using the conservity vector generated from the function
conservityTable.
}
\examples{

# Example 1: Plot a conservity against of each position for testSeqShort

data(testSeqShort)
testSeq <- testSeqShort

# Use conservityTable() to create table for testSeqShort,
# and pass it to plotOverall() for plotting
table  <- conservityTable(testSeq)
oPar <- par(mar = c(2, 4, 2, 1)) # Set plot parameters to layout plot in center.
plotOverall(table)

# Example 2: Plot a conservity against of each position for raw data
# sampleSeq3.txt. This should takes longer time to run.
findpath <- system.file("extdata", "sampleSeq3.txt", package = "conservedPos")
rawSeq <- Biostrings::readBStringSet(findpath,"fasta")

# Use conservityTable() to create table for rawSeq,
# and pass it to plotOverall() for plotting
table <- conservityTable(rawSeq)
oPar <- par(mar = c(2, 4, 2, 1)) # Set plot parameters to layout plot in center.
plotOverall(table)

}
\references{
R - Bar Charts: Bar:Chart Labels, Title and Colors.
\emph{Tutorialpoints-R Tutorial} \href{https://www.tutorialspoint.com/r/r_bar_charts.htm}{Link}

Silva, A. (2020) TestingPackage: An Example R Package For BCB410H. Unpublished.
\href{https://github.com/anjalisilva/TestingPackage}{Link}

R Bar Plot (2018) DataMentor
\emph{Getting Started in Data Science With R } \href{https://www.datamentor.io/r-programming/bar-plot/}{Link}
}
