% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyConservedPos.R
\name{runconservedPos}
\alias{runconservedPos}
\title{Launch Shiny App for conservedPos}
\usage{
runconservedPos()
}
\description{
A function runs shiny app for conservedPos
}
\examples{
\dontrun{

conservedPos::runconservedPos()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
