% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testSeqUnequal}
\alias{testSeqUnequal}
\title{A set of sequences for testing}
\format{
DNAStringSet object of length 3:
\describe{
  \item{width}{length of the sequences}
  \item{seq}{sequences}
  ...
}
}
\source{
\url{https://blast.ncbi.nlm.nih.gov/Blast.cgi}
}
\usage{
testSeqUnequal
}
\description{
A DNAStringSet containing several short and long nucleotide sequences
}
\keyword{datasets}
