% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionCal.R
\name{conservityTable}
\alias{conservityTable}
\title{This function calculates the conservity of all positions for input
sequences and generate a vector contains each position's conservity}
\usage{
conservityTable(seqSet)
}
\arguments{
\item{seqSet}{seqSet A set of aligned sequence read from fasta file}
}
\value{
a vector contains each position's highest frequency
}
\description{
This function calculates the conservity of all positions for input
sequences and generate a vector contains each position's conservity
}
\examples{
# Example 1 : Create conservity Table for testSeqShort.
data(testSeqShort)
testSeq <- testSeqShort
table  <- conservityTable(testSeq)
# You will see a vector containing 37 numbers (max length), these numbers are
# the highest frequency for each position.
table

# Example 2: Create conservity Table for another sequence set, a raw data
# sampleSeq3.txt. This should takes longer time to run.
findpath <- system.file("extdata", "sampleSeq3.txt", package = "conservedPos")
rawSeq <- Biostrings::readBStringSet(findpath,"fasta")
table <- conservityTable(rawSeq)
# Should have 140 numbers(max length of rawSeq), and
# each number is the the highest frequency for each position
table

}
