% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionCal.R
\name{findConservedNul}
\alias{findConservedNul}
\title{This function finds the nucleotide that has highest frequency of
in input string.}
\usage{
findConservedNul(str)
}
\arguments{
\item{str}{A string containing character nucleotides}
}
\value{
Return a vector that has one or more nucleotides
         with highest frequency
}
\description{
This function finds the nucleotide that has highest frequency of
in input string.
}
\examples{
# Example 1: find the the letter in string
# with highest occurrence.

str <- "ATTTGGGGCC"
findConservedNul(str) # "G"

# Example 2: find the highest frequency nucleotide from a
# string of nucleotides generated by createPosVec()

data(testSeqShort)
testSeq <- testSeqShort
position1Vec <- createPosVec(testSeq, 1)
findConservedNul(position1Vec) # "T"

}
\references{
Antonis. (2018) How to find the most repeated word in a vector with R
\emph{Stack oveflow}  \href{https://stackoverflow.com/questions/48632957/how-to-find-the-most-repeated-word-in-a-vector-with-r
}{Link}
}
