% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionCal.R
\name{findConservityFromS}
\alias{findConservityFromS}
\title{This function finds the highest frequency of nucleotide in input string.}
\usage{
findConservityFromS(str)
}
\arguments{
\item{str}{A string containing character nucleotides}
}
\value{
Return a numeric value for highest frequency
}
\description{
This function finds the highest frequency of nucleotide in input string.
}
