% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionCal.R
\name{findMaxLen}
\alias{findMaxLen}
\title{This function finds the max length of sequence for input sequence set}
\usage{
findMaxLen(seqSet)
}
\arguments{
\item{seqSet}{seqSet A set of aligned sequence read from fasta file}
}
\value{
A numeric value of the max length of sequence for seqSet
}
\description{
This function finds the max length of sequence for input sequence set
}
\examples{
# Example 1: find the max length of a sequence set retrieved from raw data
findpath <- system.file("extdata", "sampleSeq3.txt", package = "conservedPos")
rawSeq <- Biostrings::readBStringSet(findpath,"fasta")
maxL <- findMaxLen(rawSeq)
maxL # 140

# Example 2: find the max length of a sequence set retrieved from data
data(testSeqShort)
testSeq <- testSeqShort
maxL <- findMaxLen(testSeq)
maxL # 37

}
