% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConser.R
\name{plotOverall}
\alias{plotOverall}
\title{Plot the conservity of the positions}
\usage{
plotOverall(conserV)
}
\arguments{
\item{conserV}{a vector with serveral number of conservity}
}
\value{
a plot to view the overall alignment's conservity
}
\description{
This function plots the overall alignment's conservity
using the conservity vector generated from the function
conservityTable.
}
