% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConser.R
\name{plotOverall}
\alias{plotOverall}
\title{Plot the conservity of the positions}
\usage{
plotOverall(conserV)
}
\arguments{
\item{conserV}{a vector with serveral number of conservity}
}
\value{
a plot to view the overall alignment's conservity
}
\description{
This function plots the overall alignment's conservity
using the conservity vector generated from the function
conservityTable.
}
\examples{

# Example 1: Plot a conservity against of each position for testSeqShort

data(testSeqShort)
testSeq <- testSeqShort

# Use conservityTable() to create table for testSeqShort,
# and pass it to plotOverall() for plotting
table  <- conservityTable(testSeq)
oPar <- par(mar = c(2, 4, 2, 1)) # Set plot parameters to layout plot in center.
plotOverall(table)

# Example 2: Plot a conservity against of each position for raw data
# sampleSeq3.txt. This should takes longer time to run.
findpath <- system.file("extdata", "sampleSeq3.txt", package = "conservedPos")
rawSeq <- Biostrings::readBStringSet(findpath,"fasta")

# Use conservityTable() to create table for rawSeq,
# and pass it to plotOverall() for plotting
table <- conservityTable(rawSeq)
oPar <- par(mar = c(2, 4, 2, 1)) # Set plot parameters to layout plot in center.
plotOverall(table)

}
