% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConser.R
\name{plotPartial}
\alias{plotPartial}
\title{This function plots the conservity for selected alignment region
from user using the conservity vector.}
\usage{
plotPartial(conserV, firstIndex, lastIndex)
}
\description{
This function plots the conservity for selected alignment region
from user using the conservity vector.
}
